# Conflicts:
#	libraries/chain/apply_context.cpp
#	libraries/chain/wasm_interface.cpp
#	libraries/fc
#	libraries/fc/include/fc/crypto/webauthn_json
#	libraries/fc/secp256k1/upstream
#	pipeline.jsonc
#	plugins/mongo_db_plugin/mongo_db_plugin.cpp
#	plugins/state_history_plugin/state_history_plugin.cpp
#	programs/cleos/main.cpp
#	tests/CMakeLists.txt

+++ libraries/chain/authorization_manager.cpp
@@ -138,7 +138,7 @@ namespace eosio { namespace chain {
                                                                     )
    {
       for(const key_weight& k: auth.keys)
-         EOS_ASSERT(k.key.which() < _db.get<protocol_state_object>().num_supported_key_types, unactivated_key_type,
+         EOS_ASSERT(static_cast<uint32_t>(k.key.which()) < _db.get<protocol_state_object>().num_supported_key_types, unactivated_key_type,
            "Unactivated key type used when creating permission");
 
       auto creation_time = initial_creation_time;
@@ -177,7 +177,7 @@ namespace eosio { namespace chain {
                                                                     )
    {
       for(const key_weight& k: auth.keys)
-         EOS_ASSERT(k.key.which() < _db.get<protocol_state_object>().num_supported_key_types, unactivated_key_type,
+         EOS_ASSERT(static_cast<uint32_t>(k.key.which()) < _db.get<protocol_state_object>().num_supported_key_types, unactivated_key_type,
            "Unactivated key type used when creating permission");
 
       auto creation_time = initial_creation_time;
@@ -209,7 +209,7 @@ namespace eosio { namespace chain {
 
    void authorization_manager::modify_permission( const permission_object& permission, const authority& auth, uint32_t action_id ) {
       for(const key_weight& k: auth.keys)
-         EOS_ASSERT(k.key.which() < _db.get<protocol_state_object>().num_supported_key_types, unactivated_key_type,
+         EOS_ASSERT(static_cast<uint32_t>(k.key.which()) < _db.get<protocol_state_object>().num_supported_key_types, unactivated_key_type,
            "Unactivated key type used when modifying permission");
 
       _db.modify( permission, [&](permission_object& po) {
